projects/project_54/src/neural-backed-decision-trees/requirements.txt
pytorchcv
torchvision
nltk
scikit-learn
networkx
pytest
opencv-python==4.4.0.42
